Create Security Log Logon-Logoff Event Reporter
This script reads the security log, then displays a chronological record of local and remote logon and logoff activities, including failed attempts if enabled in Group/Local Policy.  It allows the input of a date range and a remote hostname if desired.  It also allows for output type to be specified as Table, HTML, or Grid View.  If none of these are desired, it returns the object in default list view.  Output manipulators can be used against the returned object, such as the Format- cmdlets.